About Me Added
Why I am doing this:
When I was seven years old, I disobeyed my mother by repeatedly running
beyond the bounds she allowed me to play tag in with another girl.  I
was playing tag on the sidewalk of a public parking lot and I was only
permitted to run in front of the glass of the karate studio while I
waited for my sister’s class to be over.  My mother became very angry
with me that day and her anger was the only emotion I remember from her.
As an adult, I now realize that my mother was genuinely frightened for
my safety.  When we got home, my mother brought me into the kitchen and
explained to me what could happen to a little girl if she went with or
was taken by a stranger.  I gulped glass after glass of water as she
explained how a stranger could put his penis inside me when I was not
ready for it, and how it would hurt me.  For the rest of my childhood, I
did not run outside of my mother’s approved boundaries again.
The explicit and implicit threat of sexual violence influenced and
continues to influence my behavior and thoughts in my everyday life, in
ways that I still am not fully aware of.  I know this is true for many
individuals, particularly women, throughout the world and I would like
to do what I can to try to understand, speak out against, and to end the
violence and pain.  My background knowledge regarding trauma is
primarily centered on sex crimes, but I will endeavor to broaden my
depth of knowledge so that I can intelligently discuss many different
types of trauma and violence.